Change BMM params, refactoring, updates
* Set mainchain BMM / WT^ verification params which require mainchain
connection to true by default

* Add 'CheckMainchainConnection' & 'SetNetworkActive' functions to
validation.h. CheckMainchainConnection checks the connection to the
mainchain the same way the sidechain page was. 'SetNetworkActive' is a
function that accesses the same function of g_connman but with
additional logging.

* Refactor sidechain page to use new functions

* Make 'refreshbmm' RPC command also check mainchain connection and
throw error if not connected

* Check for a mainchain connection in 'ReadBlock' 'CheckBlock' &
'CheckBlockHeader' so that sidechain daemons will also check for
mainchain connection. Previously only the Qt GUI version was doing this
correctly.