Change DoS score for BMM failure, move assertion
* Change the DoS score for BMM verification failure of new header /
block from 100 to 1. In rare cases, the sidechain may receive a new
block before the local mainchain node it is connected to receives the
block with the BMM proof. Changing the DoS ban score from 100 to 1
allows some time for the local mainchain node to sync before banning
the sending sidechain node.

* Move AssertLockNotHeld assertion in ProcessNewBlock to after updating
mainchain block cache. Updating the mainchain block cache locks CS_MAIN.
It is released before the end of the function anyway but the assertion
is pointless if we don't call it after any code with a lock.