Add a new amount field widget
Add a new amount field widget to replace the existing
BitcoinAmountField, simply called AmountField.

This newer version has more simple validation and a lot less code.

The style is the biggest difference though, the new version has a built
in unit display, no unit selection combo box, and no up/down arrows.
Instead of building on a QAbstractSpinBox the widget is just a QLineEdit
and a QLabel.

The new version retains validation / red highlighting of field when
invalid.